tech(singlestat): convert singlestat panel to typescript
